pytest: Rename pandas-specific pytest file
These tests are specific to the newly coined pandas engine, so the file
deserves a new name.
